#76: created domains/ folder; added refactored Domain class; added FormEnvelopes
