+++ domains/Domain.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Domains;
+
+interface Domain {
+
+    /**
+     * Returns domain singular and plural forms
+     *
+     * @return string[]
+     */
+    public function forms(): array;
+}
