+++ domains/Singular.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Domains;
+
+/**
+ * Domain name in singular form
+ */
+class Singular extends FormEnvelope
+{
+
+    /**
+     * {@inheritDoc}
+     */
+    public function __construct(Domain $domain)
+    {
+        parent::__construct($domain, "singular");
+    }
+}
